+++ astropy/io/fits/hdu/hdulist.py
@@ -114,6 +114,12 @@ def fitsopen(name, mode='readonly', memmap=None, save_backup=False,
             If `True`, image data is not scaled using BSCALE/BZERO values
             when read.
 
+        - **character_as_bytes** : bool
+
+            Whether to return bytes for string columns. By default this is `False`
+            and (unicode) strings are returned, but this does not respect memory
+            mapping and loads the whole column in memory when accessed.
+
         - **ignore_blank** : bool
 
             If `True`, the BLANK keyword is ignored if present.