+++ astropy/io/fits/hdu/table.py
@@ -247,6 +247,10 @@ class _TableBaseHDU(ExtensionHDU, _TableLikeHDU):
         If not given or None, it defaults to the value of the ``EXTVER``
         card of the ``header`` or 1.
         (default: None)
+    character_as_bytes : bool
+        Whether to return bytes for string columns. By default this is `False`
+        and (unicode) strings are returned, but this does not respect memory
+        mapping and loads the whole column in memory when accessed.
     """
 
     _manages_own_heap = False
@@ -259,12 +263,17 @@ class _TableBaseHDU(ExtensionHDU, _TableLikeHDU):
     which perform their own heap maintenance.
     """
 
-    def __init__(self, data=None, header=None, name=None, uint=False, ver=None):
+    def __init__(self, data=None, header=None, name=None, uint=False, ver=None,
+                 character_as_bytes=False):
+
         super().__init__(data=data, header=header, name=name, ver=ver)
 
         if header is not None and not isinstance(header, Header):
             raise ValueError('header must be a Header object.')
+
         self._uint = uint
+        self._character_as_bytes = character_as_bytes
+
         if data is DELAYED:
             # this should never happen
             if header is None:
@@ -400,6 +409,7 @@ class _TableBaseHDU(ExtensionHDU, _TableLikeHDU):
     def data(self):
         data = self._get_tbdata()
         data._coldefs = self.columns
+        data._character_as_bytes = self._character_as_bytes
         # Columns should now just return a reference to the data._coldefs
         del self.columns
         return data
@@ -701,6 +711,10 @@ class TableHDU(_TableBaseHDU):
         If not given or None, it defaults to the value of the ``EXTVER``
         card of the ``header`` or 1.
         (default: None)
+    character_as_bytes : bool
+        Whether to return bytes for string columns. By default this is `False`
+        and (unicode) strings are returned, but this does not respect memory
+        mapping and loads the whole column in memory when accessed.
 
     """
 
@@ -713,8 +727,8 @@ class TableHDU(_TableBaseHDU):
     __format_RE = re.compile(
         r'(?P<code>[ADEFIJ])(?P<width>\d+)(?:\.(?P<prec>\d+))?')
 
-    def __init__(self, data=None, header=None, name=None, ver=None):
-        super().__init__(data, header, name=name, ver=ver)
+    def __init__(self, data=None, header=None, name=None, ver=None, character_as_bytes=False):
+        super().__init__(data, header, name=name, ver=ver, character_as_bytes=character_as_bytes)
 
     @classmethod
     def match_header(cls, header):
@@ -813,13 +827,18 @@ class BinTableHDU(_TableBaseHDU):
         If not given or None, it defaults to the value of the ``EXTVER``
         card of the ``header`` or 1.
         (default: None)
+    character_as_bytes : bool
+        Whether to return bytes for string columns. By default this is `False`
+        and (unicode) strings are returned, but this does not respect memory
+        mapping and loads the whole column in memory when accessed.
 
     """
 
     _extension = 'BINTABLE'
     _ext_comment = 'binary table extension'
 
-    def __init__(self, data=None, header=None, name=None, uint=False, ver=None):
+    def __init__(self, data=None, header=None, name=None, uint=False, ver=None,
+                 character_as_bytes=False):
         from ....table import Table
         if isinstance(data, Table):
             from ..convenience import table_to_hdu
@@ -829,7 +848,8 @@ class BinTableHDU(_TableBaseHDU):
             data = hdu.data
             header = hdu.header
 
-        super().__init__(data, header, name=name, uint=uint, ver=ver)
+        super().__init__(data, header, name=name, uint=uint, ver=ver,
+                         character_as_bytes=character_as_bytes)
 
     @classmethod
     def match_header(cls, header):